+++ lib/ui/_widgets/octo_cat.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:gitdone/utility/github_api_handler.dart';
-import 'package:gitdone/utility/token_handler.dart';
+import 'package:gitdone/core/github_api_handler.dart';
+import 'package:gitdone/core/models/token_handler.dart';
 import 'package:http/http.dart';
 import 'package:provider/provider.dart';
