+++ lib/ui/_widgets/text_link.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:gitdone/scheme/app_color.dart';
+import 'package:gitdone/core/theme/app_color.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class TextLinkWidget {
