+++ lib/ui/login/oauth/login_oauth_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:gitdone/ui/view_models/login_github_view_model.dart';
-import 'package:gitdone/ui/views/login_github_view.dart';
+import 'package:gitdone/ui/login/oauth/login_oauth_view.dart';
+import 'package:gitdone/ui/login/oauth/login_oauth_view_model.dart';
 import 'package:provider/provider.dart';
 
 class LoginGithubScreen extends StatelessWidget {
