+++ test/widget_test.dart
@@ -1,8 +1,8 @@
 // This is a basic Flutter widget test.
 //
 // To perform an interaction with a widget in your test, use the WidgetTester
-// utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
+// core in the flutter_test package. For example, you can send tap and scroll
+// gestures. You can also use WidgetTester to find child _widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
 void main() {
